+++ android/src/main/kotlin/com/example/easy_video_editor/handler/CommandHandler.kt
@@ -14,6 +14,7 @@ enum class MethodName(val methodName: String) {
     CROP_VIDEO("cropVideo"),
     ROTATE_VIDEO("rotateVideo"),
     GENERATE_THUMBNAIL("generateThumbnail"),
+    GET_FRAME("getFrame"),
     COMPRESS_VIDEO("compressVideo"),
     CANCEL_OPERATION("cancelOperation"),
     GET_VIDEO_METADATA("getVideoMetadata"),
@@ -36,6 +37,7 @@ class CommandHandler (private val context: Context) {
         MethodName.CROP_VIDEO to CropVideoCommand(context),
         MethodName.ROTATE_VIDEO to RotateVideoCommand(context),
         MethodName.GENERATE_THUMBNAIL to GenerateThumbnailCommand(context),
+        MethodName.GET_FRAME to GetFrameCommand(context),
         MethodName.COMPRESS_VIDEO to CompressVideoCommand(context),
         MethodName.CANCEL_OPERATION to CancelOperationCommand(),
         MethodName.GET_VIDEO_METADATA to GetVideoMetadataCommand(context),
@@ -50,4 +52,4 @@ class CommandHandler (private val context: Context) {
         }
         handlers[methodName]?.execute(call, result) ?: result.notImplemented()
     }
-} 
+}