+++ lib/src/models/models.dart
@@ -1,2 +1,3 @@
 export 'video_operation.dart';
 export 'video_metadata.dart';
+export 'video_frame.dart';